+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 3.6.0 (YYYY-MM-DD)
+
+### Enhancements
+
+* [ObjectServer] Added `SyncSession.uploadAllLocalChanges()`.
+* Added `Nullable` annotation to methods that may return `null` in order to improve Kotlin usability. This also introduced a dependency to `com.google.code.findbugs:jsr305`.
+
+### Bug Fixes
+
+### Internal
+
+
 ## 3.5.1 (YYYY-MM-DD)
 
 ### Bug Fixes